Create ModDataLLParser.java
This class can be used in an Android project to find all the children of a specified layout object.
For example, if there was a table (TableLayout) which had contents updated by a call to a server on the internet,
the call to the server can be initiated within ModDataLLParser (since it implements Runnable and runs as
a thread).  After it completes the server call, the new data will need to be displayed.  To force the 
table's 'children cells' to be refreshed, this class will parse through all the children of the table's root
layout class looking for children of type TextView.  Hence, the call to this class can be made in an OnCreate activity after the View has
been initiated (after the setContentView).  The call code would look something like:

TableLayout mTableView = (TableLayout) findViewById(R.id.<the id of the root layout element>);
new Thread(new ModDataLLParser(mTableView)).start();